Moving Get to sqllog
