+++ pkg/kine/sqllog/logstructured.go
@@ -1,81 +1,5 @@
 package sqllog
 
-import (
-	"context"
-	"fmt"
-
-	"github.com/canonical/k8s-dqlite/pkg/kine/server"
-	"github.com/sirupsen/logrus"
-	"go.opentelemetry.io/otel/attribute"
-)
-
-type Log interface {
-	Start(ctx context.Context) error
-	Wait()
-	CurrentRevision(ctx context.Context) (int64, error)
-	List(ctx context.Context, prefix, startKey string, limit, revision int64, includeDeletes bool) (int64, []*server.Event, error)
-	Create(ctx context.Context, key string, value []byte, lease int64) (rev int64, created bool, err error)
-	Update(ctx context.Context, key string, value []byte, revision, lease int64) (rev int64, updated bool, err error)
-	Delete(ctx context.Context, key string, revision int64) (rev int64, deleted bool, err error)
-	After(ctx context.Context, prefix string, revision, limit int64) (int64, []*server.Event, error)
-	Watch(ctx context.Context, prefix string) <-chan []*server.Event
-	Count(ctx context.Context, prefix, startKey string, revision int64) (int64, int64, error)
-	DbSize(ctx context.Context) (int64, error)
-	DoCompact(ctx context.Context) error
-}
-
 type LogStructured struct {
 	*SQLLog
-}
-
-func (l *LogStructured) Get(ctx context.Context, key, rangeEnd string, limit, revision int64) (revRet int64, kvRet *server.KeyValue, errRet error) {
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.Get", otelName))
-	span.SetAttributes(
-		attribute.String("key", key),
-		attribute.String("rangeEnd", rangeEnd),
-		attribute.Int64("limit", limit),
-		attribute.Int64("revision", revision),
-	)
-	defer func() {
-		logrus.Debugf("GET %s, rev=%d => rev=%d, kv=%v, err=%v", key, revision, revRet, kvRet != nil, errRet)
-		span.SetAttributes(attribute.Int64("current-revision", revRet))
-		span.RecordError(errRet)
-		span.End()
-	}()
-
-	rev, kv, err := l.get(ctx, key, rangeEnd, limit, revision)
-	if kv == nil {
-		return rev, nil, err
-	}
-	return rev, kv, err
-}
-
-func (l *LogStructured) get(ctx context.Context, key, rangeEnd string, limit, revision int64) (int64, *server.KeyValue, error) {
-	var err error
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.get", otelName))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-	span.SetAttributes(
-		attribute.String("key", key),
-		attribute.String("rangeEnd", rangeEnd),
-		attribute.Int64("limit", limit),
-		attribute.Int64("revision", revision),
-	)
-	rev, events, err := l.SQLLog.List(ctx, key, rangeEnd, limit, revision)
-	if err == server.ErrCompacted {
-		span.AddEvent("key already compacted")
-		// ignore compacted when getting by revision
-		err = nil
-	} else if err != nil {
-		return 0, nil, err
-	}
-	if revision != 0 {
-		rev = revision
-	}
-	if len(events) == 0 {
-		return rev, nil, nil
-	}
-	return rev, events[0], nil
 }
